refactor: Create `NewBreedSheet.kt` file
- MOved the `NewBreedSheetContent` composable and helper methods into a dedicated file
